업적 UI 추가
- 업적 완료 시 업적 버튼 상태 변경 필요
- 업적 완료 시 보상 받는 프로세스 필요
- 보상 후 disable 추가 필요
- 업적 완료 시 알림 팝업 추가 필요

+++ Assets/Scripts/UIAchievementInfo.cs
@@ -19,7 +19,7 @@ public class UIAchievementInfo : MonoBehaviour {
 
         Quest.Progress progress = data.GetProgress();
         progressText.text = progress.progress.ToString() + "/" + progress.goal.ToString();
-        progressImage.transform.localScale = new Vector3(progress.progress / progress.goal, 1.0f, 1.0f);
+		progressImage.transform.localScale = new Vector3((float)progress.progress / (float)progress.goal, 1.0f, 1.0f);
 
         if (Quest.State.Complete == data.state)
         {
